package change remove view(move view.capture to capture)
